registered all models in admin panel
