+++ borrowings/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
 
-# Register your models here.
+from borrowings.models import Borrowing
+
+admin.site.register(Borrowing)
